Fix parent saving issue that was plaguing Rails 4
Citier does this silly thing where it sets the changed value of the
parents +id+ attribute to zero in attempt to force it to save, even
if none of it's attributes have really been changed.

In versions prior to Rails 4, this was fine, as changes to the +id+
attribute were ignored when generating the UPDATE query.

By setting it to nil, it seems ActiveRecord still things the record
is dirty, but doesn't generate an UPDATE statement WHERE "id"=0